Create useShowPublishingStatusDialog hook #6049
